requirements.txt
# Core dependencies
dspy-ai>=2.6.12
langchain>=0.1.0
openai>=1.3.0
transformers>=4.49.0
torch>=2.1.0
streamlit>=1.43.0
plotly>=6.0.0

# Knowledge Base and Retrieval
sentence-transformers>=3.4.1
networkx>=3.2.1
scikit-learn>=1.3.0
llama-index>=0.12.22
faiss-cpu>=1.7.4

# Text Processing
spacy>=3.8.4
textstat>=0.7.5
nltk>=3.8.1

# Data Processing
pandas>=2.2.3
numpy>=1.26.4
tqdm>=4.67.1

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0

# Utilities
python-dotenv>=1.0.1
psutil>=6.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Optional: Alternative Models
ollama>=0.1.5  # For local model support 